Separate classes into their own file
Previously we were defining multiple classes in a file and we
believe that this caused issues with rails auto loading.
