+++ app/models/loader/orchestrate.rb
@@ -380,79 +380,4 @@ module Loader
       Sequel::Model.db
     end
   end
-
-  # Responsible for creating policy. Called when a POST request is received
-  class CreatePolicy
-    def initialize(loader)
-      @loader = loader
-    end
-
-    def self.from_policy(policy_version)
-      CreatePolicy.new(Loader::Orchestrate.new(policy_version))
-    end
-
-    def call
-      @loader.setup_db_for_new_policy
-
-      @loader.delete_shadowed_and_duplicate_rows
-
-      @loader.store_policy_in_db
-    end
-
-    def new_roles
-      @loader.new_roles
-    end
-  end
-
-  # Responsible for replacing policy. Called when a PUT request is received
-  class ReplacePolicy
-    def initialize(loader)
-      @loader = loader
-    end
-
-    def self.from_policy(policy_version)
-      ReplacePolicy.new(Loader::Orchestrate.new(policy_version))
-    end
-
-    def call
-      @loader.setup_db_for_new_policy
-
-      @loader.delete_removed
-      
-      @loader.delete_shadowed_and_duplicate_rows
-
-      @loader.update_changed
-
-      @loader.store_policy_in_db
-    end
-
-    def new_roles
-      @loader.new_roles
-    end
-  end
-
-  # Responsible for modifying policy. Called when a PATCH request is received
-  class ModifyPolicy
-    def initialize(loader)
-      @loader = loader
-    end
-
-    def self.from_policy(policy_version)
-      ModifyPolicy.new(Loader::Orchestrate.new(policy_version))
-    end
-
-    def call
-      @loader.setup_db_for_new_policy
-      
-      @loader.delete_shadowed_and_duplicate_rows
-
-      @loader.update_changed
-
-      @loader.store_policy_in_db
-    end
-
-    def new_roles
-      @loader.new_roles
-    end
-  end
 end
